Split server implementation into multiple files
